requirements.txt
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.3
dill==0.3.6
Eel==0.15.0
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
gitdb==4.0.10
GitPython==3.1.29
greenlet==2.0.1
idna==3.4
lxml==4.9.1
multiprocess==0.70.14
pathos==0.3.0
pbr==5.11.0
pox==0.3.2
ppft==1.7.6.6
pycparser==2.21
PyExecJS==1.5.1
pyparsing==3.0.9
python-docx==0.8.11
PyYAML==6.0
requests==2.28.1
six==1.16.0
smmap==5.0.0
stevedore==4.1.1
tornado==6.2
translators==5.5.0
ua-parser==0.16.1
urllib3==1.26.13
user-agents==2.2.0
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.5.2
